+++ src/app/app.module.ts
@@ -81,7 +81,6 @@ import { CondicaoPagamentoService } from './sap/service/condicao-pagamento.servi
 import { SalesPersonSearchComponent } from './sap/components/search/sales-person-search/sales-person-search.component';
 import { CurrencyInputComponent } from './shared/components/currency/currency-input.component';
 import { LoginComponent } from './shared/components/login/login.component';
-import { AuthService } from './shared/service/auth.service';
 import { ListComponent } from './sap/components/marketing-document/core/list/list.component';
 import { CotacoesStatementComponent } from './sap/components/marketing-document/cotacao-statement/cotacoes-statement.component';
 import { DocumentCoreComponent } from './sap/components/marketing-document/core/document-core.component';
@@ -106,7 +105,6 @@ import { QuotationService } from './sap/service/document/quotation.service';
 import { TrocaComponent } from './sap/components/venda-futura/troca/troca.component';
 import { DescontoComponent } from './sap/components/document/desconto/desconto.component';
 import { CalculadoraModule } from './calculadora-preco-venda/calculadora.module';
-import { ParameterService } from './sap/service/parameter.service';
 import { GerarPdfComponent } from './sap/components/venda-futura/gerar-pdf/gerar-pdf.component';
 
 
@@ -217,7 +215,6 @@ const httpLoaderFactory = (http: HttpClient): TranslateHttpLoader =>  new Transl
     CityService,
     BusinessPartnerService,
     OneTimePasswordService,
-    ParameterService,
     DownPaymentService,
     BsModalService,
     ItemService,
@@ -227,7 +224,6 @@ const httpLoaderFactory = (http: HttpClient): TranslateHttpLoader =>  new Transl
     CondicaoPagamentoService,
     CotacaoService,
     PedidosVendaService,
-    AuthService,
     CurrencyPipe,
     FutureDeliverySalesService,
     { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
