+++ src/app/calculadora-preco-venda/calculadora.module.ts
@@ -1,16 +1,20 @@
 import { NgModule } from '@angular/core';
-import { CustoMecadoriaStatementComponent } from './custo-mercadoria/custo-mercadoria.component';
+import { CustoMecadoriaStatementComponent } from './components/custo-mercadoria/custo-mercadoria.component';
 import { SharedModule } from '../shared/shared.module';
-import { FormacaoPrecoStatementComponent } from './formacao/formacao.component';
+import { FormacaoPrecoStatementComponent } from './components/formacao/formacao.component';
+import { SelecaoProdutoComponent } from './components/selecao-produto/selecao-produto.component';
+import { CalculadoraStatementComponent } from './components/statement/statement.component';
 
 
 @NgModule({
   declarations: [
     CustoMecadoriaStatementComponent,
     FormacaoPrecoStatementComponent,
+    SelecaoProdutoComponent,
+    CalculadoraStatementComponent,
   ],
   imports: [
-    SharedModule,
+    SharedModule
   ],
   providers: [
