+++ src/app/calculadora-preco-venda/components/selecao-produto/selecao-produto.component.html
@@ -1,0 +1,23 @@
+
+<br>
+<div class="card card-primary">
+    <div class="card-header">
+        <h3 class="card-title">Selecione</h3>
+    </div>
+    <div class="card-body">
+        <div class="row">
+            Botao para buscar produtos
+        </div>
+        <div class="row">
+            lista de produtos selecionados
+        </div>
+        <div class="row">
+            <div class="form-group">
+                <button (click)="selecionaTodosProdutos()" class="btn btn-primary">Selecionar Todos Produtos</button>
+            </div>
+        </div>
+    </div>
+    <div *ngIf="false" class="overlay">
+        <i class="fas fa-2x fa-sync-alt fa-spin"></i>
+    </div>
+</div>
