+++ src/app/calculadora-preco-venda/components/statement/statement.component.ts
@@ -1,0 +1,31 @@
+import { Component, OnInit } from '@angular/core';
+import { Analise } from '../../models/analise';
+
+@Component({
+  selector: 'statement-calc',
+  templateUrl: './statement.component.html',
+})
+export class CalculadoraStatementComponent implements OnInit {
+  
+
+  analiseSelected : Analise = null
+
+  ngOnInit(): void {
+
+  }
+
+  criarAnalise(){
+    this.analiseSelected = new Analise()
+  }
+
+  isSelectProduto(){
+    return this.analiseSelected?.produtos.length == 0
+  }
+
+  selecaoProdutos($event){
+    this.analiseSelected.produtos = $event
+  }
+
+}
+
+
