+++ src/app/calculadora-preco-venda/models/analise.ts
@@ -1,0 +1,7 @@
+import { Produto } from "./produto";
+
+export class Analise{
+
+    produtos : Array<Produto> = new Array()
+    
+}
