+++ src/app/core/core.module.ts
@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ConfigService } from './services/config.service';
 import { HttpClientModule } from '@angular/common/http';
-import { AuthService } from '../shared/service/auth.service';
 
 @NgModule({
   declarations: [],
