+++ src/app/sap/components/venda-futura/venda-futura-statement.component.ts
@@ -7,7 +7,7 @@ import { VendaFuturaService } from '../../service/venda-futura.service';
 import { VendaFutura } from '../../model/venda/venda-futura';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Subscription } from 'rxjs';
-import { ParameterService } from '../../service/parameter.service';
+import { ParameterService } from '../../../shared/service/parameter.service';
